+++ Tests/CSharp/MemberTests/MemberTests.cs
@@ -7,84 +7,6 @@ namespace ICSharpCode.CodeConverter.Tests.CSharp.MemberTests;
 public class MemberTests : ConverterTestBase
 {
     [Fact]
-    public async Task TestFieldAsync()
-    {
-        await TestConversionVisualBasicToCSharpAsync(
-            @"Class TestClass
-    Const answer As Integer = 42
-    Private value As Integer = 10
-    ReadOnly v As Integer = 15
-End Class", @"
-internal partial class TestClass
-{
-    private const int answer = 42;
-    private int value = 10;
-    private readonly int v = 15;
-}");
-    }
-
-    [Fact]
-    public async Task TestMultiArrayFieldAsync()
-    {
-        await TestConversionVisualBasicToCSharpAsync(
-            @"Class TestClass
-    Dim Parts(), Taxes(), Deposits()(), Prepaid()(), FromDate, ToDate As String
-End Class", @"
-internal partial class TestClass
-{
-    private string[] Parts, Taxes;
-    private string[][] Deposits, Prepaid;
-    private string FromDate, ToDate;
-}");
-    }
-
-    [Fact]
-    public async Task TestConstantFieldInModuleAsync()
-    {
-        await TestConversionVisualBasicToCSharpAsync(
-            @"Module TestModule
-    Const answer As Integer = 42
-End Module", @"
-internal static partial class TestModule
-{
-    private const int answer = 42;
-}");
-    }
-
-    [Fact]
-    public async Task TestConstructorVisibilityAsync()
-    {
-        await TestConversionVisualBasicToCSharpAsync(@"Class Class1
-    Sub New(x As Boolean)
-    End Sub
-End Class", @"
-internal partial class Class1
-{
-    public Class1(bool x)
-    {
-    }
-}");
-    }
-
-    [Fact]
-    public async Task TestModuleConstructorAsync()
-    {
-        await TestConversionVisualBasicToCSharpAsync(
-            @"Module Module1
-    Sub New()
-        Dim someValue As Integer = 0
-    End Sub
-End Module", @"
-internal static partial class Module1
-{
-    static Module1()
-    {
-        int someValue = 0;
-    }
-}");
-    }
-
-    [Fact]
     public async Task TestDeclareMethodVisibilityInModuleAsync()
     {
         await TestConversionVisualBasicToCSharpAsync(@"Module Module1
@@ -109,59 +31,6 @@ internal partial class Class1
 {
     [DllImport(""lib.dll"")]
     public static extern void External();
-}");
-    }
-
-    [Fact]
-    public async Task TestTypeInferredConstAsync()
-    {
-        await TestConversionVisualBasicToCSharpAsync(
-            @"Class TestClass
-    Const someConstField = 42
-    Sub TestMethod()
-        Const someConst = System.DateTimeKind.Local
-    End Sub
-End Class", @"using System;
-
-internal partial class TestClass
-{
-    private const int someConstField = 42;
-    public void TestMethod()
-    {
-        const DateTimeKind someConst = DateTimeKind.Local;
-    }
-}");
-    }
-
-    [Fact]
-    public async Task TestTypeInferredVarAsync()
-    {
-        // VB doesn't infer the type of EnumVariable like you'd think, it just uses object
-        // VB compiler uses Conversions rather than any plainer casting
-        await TestConversionVisualBasicToCSharpAsync(
-            @"Class TestClass
-    Public Enum TestEnum As Integer
-        Test1
-    End Enum
-
-    Dim EnumVariable = TestEnum.Test1
-    Public Sub AMethod()
-        Dim t1 As Integer = EnumVariable
-    End Sub
-End Class", @"using Microsoft.VisualBasic.CompilerServices; // Install-Package Microsoft.VisualBasic
-
-internal partial class TestClass
-{
-    public enum TestEnum : int
-    {
-        Test1
-    }
-
-    private object EnumVariable = TestEnum.Test1;
-    public void AMethod()
-    {
-        int t1 = Conversions.ToInteger(EnumVariable);
-    }
 }");
     }
 
@@ -527,471 +396,6 @@ internal partial class TestClass
     }
 
     [Fact]
-    public async Task TestAbstractMethodAndPropertyAsync()
-    {
-        await TestConversionVisualBasicToCSharpAsync(
-            @"MustInherit Class TestClass
-    Public MustOverride Sub TestMethod()
-    Public MustOverride ReadOnly Property AbstractProperty As String
-End Class", @"
-internal abstract partial class TestClass
-{
-    public abstract void TestMethod();
-    public abstract string AbstractProperty { get; }
-}");
-    }
-
-    [Fact]
-    public async Task TestAbstractReadOnlyAndWriteOnlyPropertyAsync()
-    {
-        await TestConversionVisualBasicToCSharpAsync(
-            @"MustInherit Class TestClass
-        Public MustOverride ReadOnly Property ReadOnlyProp As String
-        Public MustOverride WriteOnly Property WriteOnlyProp As String
-End Class
-
-Class ChildClass
-    Inherits TestClass
-
-    Public Overrides ReadOnly Property ReadOnlyProp As String
-    Public Overrides WriteOnly Property WriteOnlyProp As String
-        Set
-        End Set
-    End Property
-End Class
-", @"
-internal abstract partial class TestClass
-{
-    public abstract string ReadOnlyProp { get; }
-    public abstract string WriteOnlyProp { set; }
-}
-
-internal partial class ChildClass : TestClass
-{
-
-    public override string ReadOnlyProp { get; }
-    public override string WriteOnlyProp
-    {
-        set
-        {
-        }
-    }
-}");
-    }
-
-    [Fact]
-    public async Task TestReadOnlyOrWriteOnlyPropertyImplementedByNormalPropertyAsync()
-    {
-        await TestConversionVisualBasicToCSharpAsync(
-            @"
-Interface IClass
-    ReadOnly Property ReadOnlyPropParam(i as Integer) As Integer
-    ReadOnly Property ReadOnlyProp As Integer
-
-    WriteOnly Property WriteOnlyPropParam(i as Integer) As Integer
-    WriteOnly Property WriteOnlyProp As Integer
-End Interface
-
-Class ChildClass
-    Implements IClass
-
-    Public Overridable Property RenamedPropertyParam(i As Integer) As Integer Implements IClass.ReadOnlyPropParam
-        Get
-            Return 1
-        End Get
-        Set
-        End Set
-    End Property
-
-    Public Overridable Property RenamedReadOnlyProperty As Integer Implements IClass.ReadOnlyProp ' Comment moves because this line gets split
-        Get
-            Return 2
-        End Get
-        Set
-        End Set
-    End Property
-
-    Public Overridable Property RenamedWriteOnlyPropParam(i As Integer) As Integer Implements IClass.WriteOnlyPropParam
-        Get
-            Return 1
-        End Get
-        Set
-        End Set
-    End Property
-
-    Public Overridable Property RenamedWriteOnlyProperty As Integer Implements IClass.WriteOnlyProp ' Comment moves because this line gets split
-        Get
-            Return 2
-        End Get
-        Set
-        End Set
-    End Property
-End Class
-", @"
-internal partial interface IClass
-{
-    int get_ReadOnlyPropParam(int i);
-    int ReadOnlyProp { get; }
-
-    void set_WriteOnlyPropParam(int i, int value);
-    int WriteOnlyProp { set; }
-}
-
-internal partial class ChildClass : IClass
-{
-
-    public virtual int get_RenamedPropertyParam(int i)
-    {
-        return 1;
-    }
-    public virtual void set_RenamedPropertyParam(int i, int value)
-    {
-    }
-    int IClass.get_ReadOnlyPropParam(int i) => get_RenamedPropertyParam(i);
-
-    public virtual int RenamedReadOnlyProperty
-    {
-        get
-        {
-            return 2;
-        }
-        set
-        {
-        }
-    }
-
-    int IClass.ReadOnlyProp { get => RenamedReadOnlyProperty; } // Comment moves because this line gets split
-
-    public virtual int get_RenamedWriteOnlyPropParam(int i)
-    {
-        return 1;
-    }
-    public virtual void set_RenamedWriteOnlyPropParam(int i, int value)
-    {
-    }
-    void IClass.set_WriteOnlyPropParam(int i, int value) => set_RenamedWriteOnlyPropParam(i, value);
-
-    public virtual int RenamedWriteOnlyProperty
-    {
-        get
-        {
-            return 2;
-        }
-        set
-        {
-        }
-    }
-
-    int IClass.WriteOnlyProp { set => RenamedWriteOnlyProperty = value; } // Comment moves because this line gets split
-}");
-    }
-
-    [Fact]
-    public async Task SetterProperty1053Async()
-    {
-        await TestConversionVisualBasicToCSharpAsync(
-            @"
-Public Property Prop(ByVal i As Integer) As String
-    Get
-        Static bGet As Boolean
-        bGet = False
-    End Get
-
-    Set(ByVal s As String)
-        Static bSet As Boolean
-        bSet = False
-    End Set
-End Property
-", @"
-internal partial class SurroundingClass
-{
-    private bool _Prop_bGet;
-    private bool _Prop_bSet;
-
-    public string get_Prop(int i)
-    {
-        _Prop_bGet = false;
-        return default;
-    }
-
-    public void set_Prop(int i, string value)
-    {
-        _Prop_bSet = false;
-    }
-
-}");
-    }
-
-    [Fact]
-    public async Task StaticLocalsInPropertyGetterAndSetterAsync()
-    {
-        await TestConversionVisualBasicToCSharpAsync(
-            @"
-Public Property Prop As String
-    Get
-        Static b As Boolean
-        b = True
-    End Get
-
-    Set(ByVal s As String)
-        Static b As Boolean
-        b = False
-    End Set
-End Property
-", @"
-internal partial class SurroundingClass
-{
-    private bool _Prop_b;
-    private bool _Prop_b1;
-
-    public string Prop
-    {
-        get
-        {
-            _Prop_b = true;
-            return default;
-        }
-
-        set
-        {
-            _Prop_b1 = false;
-        }
-    }
-
-}");
-    }
-
-    [Fact]
-    public async Task TestReadOnlyAndWriteOnlyParametrizedPropertyAsync()
-    {
-        await TestConversionVisualBasicToCSharpAsync(
-            @"Interface IClass
-    ReadOnly Property ReadOnlyProp(i as Integer) As String
-    WriteOnly Property WriteOnlyProp(i as Integer) As String
-End Interface
-
-Class ChildClass
-    Implements IClass
-
-    Public Overridable ReadOnly Property ReadOnlyProp(i As Integer) As String Implements IClass.ReadOnlyProp
-        Get
-            Throw New NotImplementedException
-        End Get
-    End Property
-
-    Public Overridable WriteOnly Property WriteOnlyProp(i As Integer) As String Implements IClass.WriteOnlyProp
-        Set
-            Throw New NotImplementedException
-        End Set
-    End Property
-End Class
-", @"using System;
-
-internal partial interface IClass
-{
-    string get_ReadOnlyProp(int i);
-    void set_WriteOnlyProp(int i, string value);
-}
-
-internal partial class ChildClass : IClass
-{
-
-    public virtual string get_ReadOnlyProp(int i)
-    {
-        throw new NotImplementedException();
-    }
-
-    public virtual void set_WriteOnlyProp(int i, string value)
-    {
-        throw new NotImplementedException();
-    }
-}");
-    }
-
-    [Fact]
-    public async Task TestExplicitInterfaceOfParametrizedPropertyAsync()
-    {
-        await TestConversionVisualBasicToCSharpAsync(
-            @"Interface IClass
-    ReadOnly Property ReadOnlyPropToRename(i as Integer) As String
-    WriteOnly Property WriteOnlyPropToRename(i as Integer) As String
-    Property PropToRename(i as Integer) As String
-
-    ReadOnly Property ReadOnlyPropNonPublic(i as Integer) As String
-    WriteOnly Property WriteOnlyPropNonPublic(i as Integer) As String
-    Property PropNonPublic(i as Integer) As String
-
-    ReadOnly Property ReadOnlyPropToRenameNonPublic(i as Integer) As String
-    WriteOnly Property WriteOnlyPropToRenameNonPublic(i as Integer) As String
-    Property PropToRenameNonPublic(i as Integer) As String
-
-End Interface
-
-Class ChildClass
-    Implements IClass
-
-    Public ReadOnly Property ReadOnlyPropRenamed(i As Integer) As String Implements IClass.ReadOnlyPropToRename
-        Get
-            Throw New NotImplementedException
-        End Get
-    End Property
-
-    Public Overridable WriteOnly Property WriteOnlyPropRenamed(i As Integer) As String Implements IClass.WriteOnlyPropToRename
-        Set
-            Throw New NotImplementedException
-        End Set
-    End Property
-
-    Public Overridable Property PropRenamed(i As Integer) As String Implements IClass.PropToRename
-        Get
-            Throw New NotImplementedException
-        End Get
-        Set
-            Throw New NotImplementedException
-        End Set
-    End Property
-
-    Private ReadOnly Property ReadOnlyPropNonPublic(i As Integer) As String Implements IClass.ReadOnlyPropNonPublic
-        Get
-            Throw New NotImplementedException
-        End Get
-    End Property
-
-    Protected Friend Overridable WriteOnly Property WriteOnlyPropNonPublic(i As Integer) As String Implements IClass.WriteOnlyPropNonPublic
-        Set
-            Throw New NotImplementedException
-        End Set
-    End Property
-
-    Friend Overridable Property PropNonPublic(i As Integer) As String Implements IClass.PropNonPublic
-        Get
-            Throw New NotImplementedException
-        End Get
-        Set
-            Throw New NotImplementedException
-        End Set
-    End Property
-
-    Protected Friend Overridable ReadOnly Property ReadOnlyPropRenamedNonPublic(i As Integer) As String Implements IClass.ReadOnlyPropToRenameNonPublic
-        Get
-            Throw New NotImplementedException
-        End Get
-    End Property
-
-    Private WriteOnly Property WriteOnlyPropRenamedNonPublic(i As Integer) As String Implements IClass.WriteOnlyPropToRenameNonPublic
-        Set
-            Throw New NotImplementedException
-        End Set
-    End Property
-
-    Friend Overridable Property PropToRenameNonPublic(i As Integer) As String Implements IClass.PropToRenameNonPublic
-        Get
-            Throw New NotImplementedException
-        End Get
-        Set
-            Throw New NotImplementedException
-        End Set
-    End Property
-End Class
-", @"using System;
-
-internal partial interface IClass
-{
-    string get_ReadOnlyPropToRename(int i);
-    void set_WriteOnlyPropToRename(int i, string value);
-    string get_PropToRename(int i);
-    void set_PropToRename(int i, string value);
-
-    string get_ReadOnlyPropNonPublic(int i);
-    void set_WriteOnlyPropNonPublic(int i, string value);
-    string get_PropNonPublic(int i);
-    void set_PropNonPublic(int i, string value);
-
-    string get_ReadOnlyPropToRenameNonPublic(int i);
-    void set_WriteOnlyPropToRenameNonPublic(int i, string value);
-    string get_PropToRenameNonPublic(int i);
-    void set_PropToRenameNonPublic(int i, string value);
-
-}
-
-internal partial class ChildClass : IClass
-{
-
-    public string get_ReadOnlyPropRenamed(int i)
-    {
-        throw new NotImplementedException();
-    }
-    string IClass.get_ReadOnlyPropToRename(int i) => get_ReadOnlyPropRenamed(i);
-
-    public virtual void set_WriteOnlyPropRenamed(int i, string value)
-    {
-        throw new NotImplementedException();
-    }
-    void IClass.set_WriteOnlyPropToRename(int i, string value) => set_WriteOnlyPropRenamed(i, value);
-
-    public virtual string get_PropRenamed(int i)
-    {
-        throw new NotImplementedException();
-    }
-    public virtual void set_PropRenamed(int i, string value)
-    {
-        throw new NotImplementedException();
-    }
-
-    string IClass.get_PropToRename(int i) => get_PropRenamed(i);
-    void IClass.set_PropToRename(int i, string value) => set_PropRenamed(i, value);
-
-    private string get_ReadOnlyPropNonPublic(int i)
-    {
-        throw new NotImplementedException();
-    }
-    string IClass.get_ReadOnlyPropNonPublic(int i) => get_ReadOnlyPropNonPublic(i);
-
-    protected internal virtual void set_WriteOnlyPropNonPublic(int i, string value)
-    {
-        throw new NotImplementedException();
-    }
-    void IClass.set_WriteOnlyPropNonPublic(int i, string value) => set_WriteOnlyPropNonPublic(i, value);
-
-    internal virtual string get_PropNonPublic(int i)
-    {
-        throw new NotImplementedException();
-    }
-    internal virtual void set_PropNonPublic(int i, string value)
-    {
-        throw new NotImplementedException();
-    }
-
-    string IClass.get_PropNonPublic(int i) => get_PropNonPublic(i);
-    void IClass.set_PropNonPublic(int i, string value) => set_PropNonPublic(i, value);
-
-    protected internal virtual string get_ReadOnlyPropRenamedNonPublic(int i)
-    {
-        throw new NotImplementedException();
-    }
-    string IClass.get_ReadOnlyPropToRenameNonPublic(int i) => get_ReadOnlyPropRenamedNonPublic(i);
-
-    private void set_WriteOnlyPropRenamedNonPublic(int i, string value)
-    {
-        throw new NotImplementedException();
-    }
-    void IClass.set_WriteOnlyPropToRenameNonPublic(int i, string value) => set_WriteOnlyPropRenamedNonPublic(i, value);
-
-    internal virtual string get_PropToRenameNonPublic(int i)
-    {
-        throw new NotImplementedException();
-    }
-    internal virtual void set_PropToRenameNonPublic(int i, string value)
-    {
-        throw new NotImplementedException();
-    }
-
-    string IClass.get_PropToRenameNonPublic(int i) => get_PropToRenameNonPublic(i);
-    void IClass.set_PropToRenameNonPublic(int i, string value) => set_PropToRenameNonPublic(i, value);
-}");
-    }
-
-    [Fact]
     public async Task TestSealedMethodAsync()
     {
         await TestConversionVisualBasicToCSharpAsync(
@@ -1063,272 +467,6 @@ internal partial class TestSubclass : TestClass
     }
 
     [Fact]
-    public async Task TestExtensionMethodAsync()
-    {
-        await TestConversionVisualBasicToCSharpAsync(
-            @"Module TestClass
-    <System.Runtime.CompilerServices.Extension()>
-    Sub TestMethod(ByVal str As String)
-    End Sub
-
-    <System.Runtime.CompilerServices.Extension()>
-    Sub TestMethod2Parameters(ByVal str As String, other As String)
-    End Sub
-End Module", @"
-internal static partial class TestClass
-{
-    public static void TestMethod(this string str)
-    {
-    }
-
-    public static void TestMethod2Parameters(this string str, string other)
-    {
-    }
-}");
-    }
-
-    [Fact]
-    public async Task TestExtensionMethodWithExistingImportAsync()
-    {
-        await TestConversionVisualBasicToCSharpAsync(
-            @"Imports System.Runtime.CompilerServices ' Removed by simplifier
-
-Module TestClass
-    <Extension()>
-    Sub TestMethod(ByVal str As String)
-    End Sub
-End Module", @"
-internal static partial class TestClass
-{
-    public static void TestMethod(this string str)
-    {
-    }
-}");
-    }
-
-    [Fact]
-    public async Task TestRefExtensionMethodAsync()
-    {
-        await TestConversionVisualBasicToCSharpAsync(
-            @"Imports System
-Imports System.Runtime.CompilerServices ' Removed since the extension attribute is removed
-
-Public Module MyExtensions
-    <Extension()>
-    Public Sub Add(Of T)(ByRef arr As T(), item As T)
-        Array.Resize(arr, arr.Length + 1)
-        arr(arr.Length - 1) = item
-    End Sub
-End Module
-
-Public Module UsagePoint
-    Public Sub Main()
-        Dim arr = New Integer() {1, 2, 3}
-        arr.Add(4)
-        System.Console.WriteLine(arr(3))
-    End Sub
-End Module", @"using System;
-
-public static partial class MyExtensions
-{
-    public static void Add<T>(ref T[] arr, T item)
-    {
-        Array.Resize(ref arr, arr.Length + 1);
-        arr[arr.Length - 1] = item;
-    }
-}
-
-public static partial class UsagePoint
-{
-    public static void Main()
-    {
-        int[] arr = new int[] { 1, 2, 3 };
-        MyExtensions.Add(ref arr, 4);
-        Console.WriteLine(arr[3]);
-    }
-}");
-    }
-
-    [Fact]
-    public async Task TestExtensionWithinExtendedTypeAsync()
-    {
-        await TestConversionVisualBasicToCSharpAsync(
-            @"Module Extensions
-    <Extension()>
-    Sub TestExtension(extendedClass As ExtendedClass)
-    End Sub
-End Module
-
-Class ExtendedClass
-  Sub TestExtensionConsumer()
-    TestExtension()
-  End Sub
-End Class", @"
-internal static partial class Extensions
-{
-    public static void TestExtension(this ExtendedClass extendedClass)
-    {
-    }
-}
-
-internal partial class ExtendedClass
-{
-    public void TestExtensionConsumer()
-    {
-        this.TestExtension();
-    }
-}");
-    }
-
-    [Fact]
-    public async Task TestExtensionWithinTypeDerivedFromExtendedTypeAsync()
-    {
-        await TestConversionVisualBasicToCSharpAsync(
-            @"Module Extensions
-    <Extension()>
-    Sub TestExtension(extendedClass As ExtendedClass)
-    End Sub
-End Module
-
-Class ExtendedClass
-End Class
-
-Class DerivedClass
-    Inherits ExtendedClass
-
-  Sub TestExtensionConsumer()
-    TestExtension()
-  End Sub
-End Class", @"
-internal static partial class Extensions
-{
-    public static void TestExtension(this ExtendedClass extendedClass)
-    {
-    }
-}
-
-internal partial class ExtendedClass
-{
-}
-
-internal partial class DerivedClass : ExtendedClass
-{
-
-    public void TestExtensionConsumer()
-    {
-        this.TestExtension();
-    }
-}");
-    }
-
-    [Fact]
-    public async Task TestExtensionWithinNestedExtendedTypeAsync()
-    {
-        await TestConversionVisualBasicToCSharpAsync(
-            @"Module Extensions
-    <Extension()>
-    Sub TestExtension(extendedClass As NestingClass.ExtendedClass)
-    End Sub
-End Module
-
-Class NestingClass
-    Class ExtendedClass
-      Sub TestExtensionConsumer()
-        TestExtension()
-      End Sub        
-    End Class
-End Class", @"
-internal static partial class Extensions
-{
-    public static void TestExtension(this NestingClass.ExtendedClass extendedClass)
-    {
-    }
-}
-
-internal partial class NestingClass
-{
-    public partial class ExtendedClass
-    {
-        public void TestExtensionConsumer()
-        {
-            this.TestExtension();
-        }
-    }
-}");
-    }
-
-    [Fact]
-    public async Task TestExtensionWithMeWithinExtendedTypeAsync()
-    {
-        await TestConversionVisualBasicToCSharpAsync(
-            @"Module Extensions
-    <Extension()>
-    Sub TestExtension(extendedClass As ExtendedClass)
-    End Sub
-End Module
-
-Class ExtendedClass
-  Sub TestExtensionConsumer()
-    Me.TestExtension()
-  End Sub
-End Class", @"
-internal static partial class Extensions
-{
-    public static void TestExtension(this ExtendedClass extendedClass)
-    {
-    }
-}
-
-internal partial class ExtendedClass
-{
-    public void TestExtensionConsumer()
-    {
-        this.TestExtension();
-    }
-}");
-    }
-
-    [Fact]
-    public async Task TestConstructorAsync()
-    {
-        await TestConversionVisualBasicToCSharpAsync(
-            @"Class TestClass(Of T As {Class, New}, T2 As Structure, T3)
-    Public Sub New(<Out> ByRef argument As T, ByRef argument2 As T2, ByVal argument3 As T3)
-    End Sub
-End Class", @"
-internal partial class TestClass<T, T2, T3>
-    where T : class, new()
-    where T2 : struct
-{
-    public TestClass(out T argument, ref T2 argument2, T3 argument3)
-    {
-    }
-}
-1 target compilation errors:
-CS0177: The out parameter 'argument' must be assigned to before control leaves the current method");
-    }
-
-    [Fact]
-    public async Task TestConstructorWithImplicitPublicAccessibilityAsync()
-    {
-        await TestConversionVisualBasicToCSharpAsync(
-            @"Sub New()
-End Sub", @"public SurroundingClass()
-{
-}");
-    }
-
-    [Fact]
-    public async Task TestStaticConstructorAsync()
-    {
-        await TestConversionVisualBasicToCSharpAsync(
-            @"Shared Sub New()
-End Sub", @"static SurroundingClass()
-{
-}");
-    }
-
-    [Fact]
     public async Task TestDestructorAsync()
     {
         await TestConversionVisualBasicToCSharpAsync(
@@ -1348,7 +486,7 @@ internal partial class TestClass
     public async Task Issue681_OverloadsOverridesPropertyAsync()
     {
         await TestConversionVisualBasicToCSharpAsync(
-            @"Public Class C 
+            @"Public Class C
     Inherits B
 
     Public ReadOnly Overloads Overrides Property X()
@@ -1418,7 +556,7 @@ Friend Class TestClass2
 
     Public Overrides Sub CreateAbstractInstance()
     End Sub
-    
+
     Public Overloads Sub CreateVirtualInstance(o As Object)
     End Sub
 End Class",
@@ -1474,84 +612,6 @@ End Class", @"using System.Diagnostics;
 [DebuggerDisplay(""Hello World"")]
 internal partial class TestClass
 {
-}");
-    }
-
-    [Fact]
-    public async Task FieldWithAttributeAsync()
-    {
-        await TestConversionVisualBasicToCSharpAsync(@"Class TestClass
-    <ThreadStatic>
-    Private Shared First As Integer
-End Class", @"using System;
-
-internal partial class TestClass
-{
-    [ThreadStatic]
-    private static int First;
-}");
-    }
-
-    [Fact]
-    public async Task FieldWithNonStaticInitializerAsync()
-    {
-        await TestConversionVisualBasicToCSharpAsync(@"Public Class A
-    Private x As Integer = 2
-    Private y(x) As Integer
-End Class", @"
-public partial class A
-{
-    private int x = 2;
-    private int[] y;
-
-    public A()
-    {
-        y = new int[x + 1];
-    }
-}");
-    }
-
-    [Fact]
-    public async Task FieldWithInstanceOperationOfDifferingTypeAsync()
-    {
-        await TestConversionVisualBasicToCSharpAsync(@"Public Class DoesNotNeedConstructor
-    Private ReadOnly ClassVariable1 As New ParallelOptions With {.MaxDegreeOfParallelism = 5}
-End Class", @"using System.Threading.Tasks;
-
-public partial class DoesNotNeedConstructor
-{
-    private readonly ParallelOptions ClassVariable1 = new ParallelOptions() { MaxDegreeOfParallelism = 5 };
-}");
-    }
-
-    [Fact]
-    public async Task Issue281FieldWithNonStaticLambdaInitializerAsync()
-    {
-        await TestConversionVisualBasicToCSharpAsync(@"Imports System.IO
-
-Public Class Issue281
-    Private lambda As System.Delegate = New ErrorEventHandler(Sub(a, b) Len(0))
-    Private nonShared As System.Delegate = New ErrorEventHandler(AddressOf OnError)
-
-    Sub OnError(s As Object, e As ErrorEventArgs)
-    End Sub
-End Class", @"using System;
-using System.IO;
-using Microsoft.VisualBasic; // Install-Package Microsoft.VisualBasic
-
-public partial class Issue281
-{
-    private Delegate lambda = new ErrorEventHandler((a, b) => Strings.Len(0));
-    private Delegate nonShared;
-
-    public Issue281()
-    {
-        nonShared = new ErrorEventHandler(OnError);
-    }
-
-    public void OnError(object s, ErrorEventArgs e)
-    {
-    }
 }");
     }
 
@@ -1804,73 +864,6 @@ internal partial class AsyncCode
     }
 
     [Fact]
-    public async Task TestAsyncMethodsWithNoReturnAsync()
-    {
-        await TestConversionVisualBasicToCSharpAsync(
-            @"Friend Partial Module TaskExtensions
-    <Extension()>
-    Async Function [Then](Of T)(ByVal task As Task, ByVal f As Func(Of Task(Of T))) As Task(Of T)
-        Await task
-        Return Await f()
-    End Function
-
-    <Extension()>
-    Async Function [Then](ByVal task As Task, ByVal f As Func(Of Task)) As Task
-        Await task
-        Await f()
-    End Function
-
-    <Extension()>
-    Async Function [Then](Of T, U)(ByVal task As Task(Of T), ByVal f As Func(Of T, Task(Of U))) As Task(Of U)
-        Return Await f(Await task)
-    End Function
-
-    <Extension()>
-    Async Function [Then](Of T)(ByVal task As Task(Of T), ByVal f As Func(Of T, Task)) As Task
-        Await f(Await task)
-    End Function
-
-    <Extension()>
-    Async Function [ThenExit](Of T)(ByVal task As Task(Of T), ByVal f As Func(Of T, Task)) As Task
-        Await f(Await task)
-        Exit Function
-    End Function
-End Module", @"using System;
-using System.Threading.Tasks;
-
-internal static partial class TaskExtensions
-{
-    public async static Task<T> Then<T>(this Task task, Func<Task<T>> f)
-    {
-        await task;
-        return await f();
-    }
-
-    public async static Task Then(this Task task, Func<Task> f)
-    {
-        await task;
-        await f();
-    }
-
-    public async static Task<U> Then<T, U>(this Task<T> task, Func<T, Task<U>> f)
-    {
-        return await f(await task);
-    }
-
-    public async static Task Then<T>(this Task<T> task, Func<T, Task> f)
-    {
-        await f(await task);
-    }
-
-    public async static Task ThenExit<T>(this Task<T> task, Func<T, Task> f)
-    {
-        await f(await task);
-        return;
-    }
-}");
-    }
-
-    [Fact]
     public async Task TestExternDllImportAsync()
     {
         await TestConversionVisualBasicToCSharpAsync(
@@ -1883,525 +876,6 @@ private static extern IntPtr OpenProcess(AccessMask dwDesiredAccess, bool bInher
 BC30002: Type 'AccessMask' is not defined.
 1 target compilation errors:
 CS0246: The type or namespace name 'AccessMask' could not be found (are you missing a using directive or an assembly reference?)");
-    }
-
-    [Fact]
-    public async Task Issue443_FixCaseForInterfaceMembersAsync()
-    {
-        await TestConversionVisualBasicToCSharpAsync(
-            @"Public Interface IFoo
-    Function FooDifferentCase(<Out> ByRef str2 As String) As Integer
-End Interface
-
-Public Class Foo
-    Implements IFoo
-    Function fooDifferentCase(<Out> ByRef str2 As String) As Integer Implements IFoo.FOODIFFERENTCASE
-        str2 = 2.ToString()
-        Return 3
-    End Function
-End Class", @"
-public partial interface IFoo
-{
-    int FooDifferentCase(out string str2);
-}
-
-public partial class Foo : IFoo
-{
-    public int FooDifferentCase(out string str2)
-    {
-        str2 = 2.ToString();
-        return 3;
-    }
-}
-");
-    }
-
-    [Fact]
-    public async Task Issue444_FixNameForRenamedInterfaceMembersAsync()
-    {
-        await TestConversionVisualBasicToCSharpAsync(
-            @"Public Interface IFoo
-    Function FooDifferentName(ByRef str As String, i As Integer) As Integer
-End Interface
-
-Public Class Foo
-    Implements IFoo
-
-    Function BarDifferentName(ByRef str As String, i As Integer) As Integer Implements IFoo.FooDifferentName
-        Return 4
-    End Function
-End Class", @"
-public partial interface IFoo
-{
-    int FooDifferentName(ref string str, int i);
-}
-
-public partial class Foo : IFoo
-{
-
-    public int BarDifferentName(ref string str, int i)
-    {
-        return 4;
-    }
-
-    int IFoo.FooDifferentName(ref string str, int i) => BarDifferentName(ref str, i);
-}
-");
-    }
-
-    [Fact]
-    public async Task IdenticalInterfaceMethodsWithRenamedInterfaceMembersAsync()
-    {
-        await TestConversionVisualBasicToCSharpAsync(
-            @"Public Interface IFoo
-    Function DoFooBar(ByRef str As String, i As Integer) As Integer
-End Interface
-
-Public Interface IBar
-    Function DoFooBar(ByRef str As String, i As Integer) As Integer
-End Interface
-
-Public Class FooBar
-    Implements IFoo, IBar
-
-    Function Foo(ByRef str As String, i As Integer) As Integer Implements IFoo.DoFooBar
-        Return 4
-    End Function
-
-    Function Bar(ByRef str As String, i As Integer) As Integer Implements IBar.DoFooBar
-        Return 2
-    End Function
-
-End Class", @"
-public partial interface IFoo
-{
-    int DoFooBar(ref string str, int i);
-}
-
-public partial interface IBar
-{
-    int DoFooBar(ref string str, int i);
-}
-
-public partial class FooBar : IFoo, IBar
-{
-
-    public int Foo(ref string str, int i)
-    {
-        return 4;
-    }
-
-    int IFoo.DoFooBar(ref string str, int i) => Foo(ref str, i);
-
-    public int Bar(ref string str, int i)
-    {
-        return 2;
-    }
-
-    int IBar.DoFooBar(ref string str, int i) => Bar(ref str, i);
-
-}
-");
-    }
-
-    [Fact]
-    public async Task RenamedInterfaceCasingOnlyDifferenceConsumerAsync()
-    {
-        await TestConversionVisualBasicToCSharpAsync(
-            @"
-Public Interface IFoo
-    Function DoFoo() As Integer
-    Property Prop As Integer
-End Interface
-
-Public Class Foo
-    Implements IFoo
-
-    Private Function doFoo() As Integer Implements IFoo.DoFoo
-        Return 4
-    End Function
-
-    Private Property prop As Integer Implements IFoo.Prop
-
-    Private Function Consumer() As Integer
-        Dim foo As New Foo()
-        Dim interfaceInstance As IFoo = foo
-        Return foo.doFoo() + foo.DoFoo() +
-               interfaceInstance.doFoo() + interfaceInstance.DoFoo() +
-               foo.prop + foo.Prop +
-               interfaceInstance.prop + interfaceInstance.Prop
-    End Function
-
-End Class", @"
-public partial interface IFoo
-{
-    int DoFoo();
-    int Prop { get; set; }
-}
-
-public partial class Foo : IFoo
-{
-
-    private int doFoo()
-    {
-        return 4;
-    }
-
-    int IFoo.DoFoo() => doFoo();
-
-    private int prop { get; set; }
-    int IFoo.Prop { get => prop; set => prop = value; }
-
-    private int Consumer()
-    {
-        var foo = new Foo();
-        IFoo interfaceInstance = foo;
-        return foo.doFoo() + foo.doFoo() + interfaceInstance.DoFoo() + interfaceInstance.DoFoo() + foo.prop + foo.prop + interfaceInstance.Prop + interfaceInstance.Prop;
-    }
-
-}
-");
-    }
-
-    [Fact]
-    public async Task RenamedInterfaceCasingOnlyDifferenceForVirtualMemberConsumerAsync()
-    {
-        await TestConversionVisualBasicToCSharpAsync(
-            @"
-Public Interface IFoo
-    Function DoFoo() As Integer
-    Property Prop As Integer
-End Interface
-
-Public MustInherit Class BaseFoo
-    Implements IFoo
-
-    Protected Friend Overridable Function doFoo() As Integer Implements IFoo.DoFoo
-        Return 4
-    End Function
-
-    Protected Friend Overridable Property prop As Integer Implements IFoo.Prop
-
-End Class
-
-Public Class Foo
-    Inherits BaseFoo
-
-    Protected Friend Overrides Function DoFoo() As Integer
-        Return 5
-    End Function
-
-    Protected Friend Overrides Property Prop As Integer
-
-    Private Function Consumer() As Integer
-        Dim foo As New Foo()
-        Dim interfaceInstance As IFoo = foo
-        Dim baseClass As BaseFoo = foo
-        Return foo.doFoo() +  foo.DoFoo() +
-               interfaceInstance.doFoo() + interfaceInstance.DoFoo() + 
-               baseClass.doFoo() + baseClass.DoFoo() +
-               foo.prop + foo.Prop +
-               interfaceInstance.prop + interfaceInstance.Prop +
-               baseClass.prop + baseClass.Prop
-    End Function
-End Class", @"
-public partial interface IFoo
-{
-    int DoFoo();
-    int Prop { get; set; }
-}
-
-public abstract partial class BaseFoo : IFoo
-{
-
-    protected internal virtual int doFoo()
-    {
-        return 4;
-    }
-
-    int IFoo.DoFoo() => doFoo();
-
-    protected internal virtual int prop { get; set; }
-    int IFoo.Prop { get => prop; set => prop = value; }
-
-}
-
-public partial class Foo : BaseFoo
-{
-
-    protected internal override int doFoo()
-    {
-        return 5;
-    }
-
-    protected internal override int prop { get; set; }
-
-    private int Consumer()
-    {
-        var foo = new Foo();
-        IFoo interfaceInstance = foo;
-        BaseFoo baseClass = foo;
-        return foo.doFoo() + foo.doFoo() + interfaceInstance.DoFoo() + interfaceInstance.DoFoo() + baseClass.doFoo() + baseClass.doFoo() + foo.prop + foo.prop + interfaceInstance.Prop + interfaceInstance.Prop + baseClass.prop + baseClass.prop;
-    }
-}
-");
-    }
-
-    [Fact]
-    public async Task RenamedInterfaceCasingOnlyDifferenceWithOverloadedPropertyConsumerAsync()
-    {
-        await TestConversionVisualBasicToCSharpAsync(
-            @"
-Public Interface IUserContext
-    ReadOnly Property GroupID As String
-End Interface
-
-Public Interface IFoo
-    ReadOnly Property ConnectedGroupId As String
-End Interface
-
-Public MustInherit Class BaseFoo
-    Implements IUserContext
-
-    Protected Friend ReadOnly Property ConnectedGroupID() As String Implements IUserContext.GroupID
-
-End Class
-
-Public Class Foo
-    Inherits BaseFoo
-    Implements IFoo
-
-    Protected Friend Overloads ReadOnly Property ConnectedGroupID As String Implements IFoo.ConnectedGroupId ' Comment moves because this line gets split
-        Get
-            Return If("""", MyBase.ConnectedGroupID())
-        End Get
-    End Property
-
-    Private Function Consumer() As String
-        Dim foo As New Foo()
-        Dim ifoo As IFoo = foo
-        Dim baseFoo As BaseFoo = foo
-        Dim iUserContext As IUserContext = foo
-        Return foo.ConnectedGroupID & foo.ConnectedGroupId & 
-               iFoo.ConnectedGroupID & iFoo.ConnectedGroupId &
-               baseFoo.ConnectedGroupID & baseFoo.ConnectedGroupId &
-               iUserContext.GroupId & iUserContext.GroupID
-    End Function
-
-End Class", @"
-public partial interface IUserContext
-{
-    string GroupID { get; }
-}
-
-public partial interface IFoo
-{
-    string ConnectedGroupId { get; }
-}
-
-public abstract partial class BaseFoo : IUserContext
-{
-
-    protected internal string ConnectedGroupID { get; private set; }
-    string IUserContext.GroupID { get => ConnectedGroupID; }
-
-}
-
-public partial class Foo : BaseFoo, IFoo
-{
-
-    protected internal new string ConnectedGroupID
-    {
-        get
-        {
-            return """" ?? base.ConnectedGroupID;
-        }
-    }
-
-    string IFoo.ConnectedGroupId { get => ConnectedGroupID; } // Comment moves because this line gets split
-
-    private string Consumer()
-    {
-        var foo = new Foo();
-        IFoo ifoo = foo;
-        BaseFoo baseFoo = foo;
-        IUserContext iUserContext = foo;
-        return foo.ConnectedGroupID + foo.ConnectedGroupID + ifoo.ConnectedGroupId + ifoo.ConnectedGroupId + baseFoo.ConnectedGroupID + baseFoo.ConnectedGroupID + iUserContext.GroupID + iUserContext.GroupID;
-    }
-
-}
-");
-    }
-
-    [Fact]
-    public async Task RenamedMethodImplementsMultipleInterfacesAsync()
-    {
-        await TestConversionVisualBasicToCSharpAsync(
-            @"Public Interface IFoo
-    Function DoFooBar(ByRef str As String, i As Integer) As Integer
-End Interface
-
-Public Interface IBar
-    Function DoFooBar(ByRef str As String, i As Integer) As Integer
-End Interface
-
-Public Class FooBar
-    Implements IFoo, IBar
-
-    Function Foo(ByRef str As String, i As Integer) As Integer Implements IFoo.DoFooBar, IBar.DoFooBar
-        Return 4
-    End Function
-
-End Class", @"
-public partial interface IFoo
-{
-    int DoFooBar(ref string str, int i);
-}
-
-public partial interface IBar
-{
-    int DoFooBar(ref string str, int i);
-}
-
-public partial class FooBar : IFoo, IBar
-{
-
-    public int Foo(ref string str, int i)
-    {
-        return 4;
-    }
-
-    int IFoo.DoFooBar(ref string str, int i) => Foo(ref str, i);
-    int IBar.DoFooBar(ref string str, int i) => Foo(ref str, i);
-
-}");
-    }
-
-    [Fact]
-    public async Task IdenticalInterfacePropertiesWithRenamedInterfaceMembersAsync()
-    {
-        await TestConversionVisualBasicToCSharpAsync(
-            @"Public Interface IFoo
-        Property FooBarProp As Integer
-    End Interface
-
-Public Interface IBar
-    Property FooBarProp As Integer
-End Interface
-
-Public Class FooBar
-    Implements IFoo, IBar
-
-    Property Foo As Integer Implements IFoo.FooBarProp
-
-    Property Bar As Integer Implements IBar.FooBarProp
-    
-End Class", @"
-public partial interface IFoo
-{
-    int FooBarProp { get; set; }
-}
-
-public partial interface IBar
-{
-    int FooBarProp { get; set; }
-}
-
-public partial class FooBar : IFoo, IBar
-{
-
-    public int Foo { get; set; }
-    int IFoo.FooBarProp { get => Foo; set => Foo = value; }
-
-    public int Bar { get; set; }
-    int IBar.FooBarProp { get => Bar; set => Bar = value; }
-
-}");
-    }
-
-    [Fact]
-    public async Task ExplicitInterfaceImplementationRequiredMethodParameters_749_Async()
-    {
-        await TestConversionVisualBasicToCSharpAsync(
-            @"
-Public Interface IFoo
-  Function DoFooBar(ByRef str As String, i As Integer) As Integer
-End Interface
-
-Public Interface IBar
-  Function DoFooBar(ByRef str As String, i As Integer) As Integer
-End Interface
-
-Public Class FooBar
-  Implements IFoo, IBar
-
-  Function Foo(ByRef str As String, i As Integer) As Integer Implements IFoo.DoFooBar
-    Return 4
-  End Function
-
-  Function Bar(ByRef str As String, i As Integer) As Integer Implements IBar.DoFooBar
-    Return 2
-  End Function
-
-End Class", @"
-public partial interface IFoo
-{
-    int DoFooBar(ref string str, int i);
-}
-
-public partial interface IBar
-{
-    int DoFooBar(ref string str, int i);
-}
-
-public partial class FooBar : IFoo, IBar
-{
-
-    public int Foo(ref string str, int i)
-    {
-        return 4;
-    }
-
-    int IFoo.DoFooBar(ref string str, int i) => Foo(ref str, i);
-
-    public int Bar(ref string str, int i)
-    {
-        return 2;
-    }
-
-    int IBar.DoFooBar(ref string str, int i) => Bar(ref str, i);
-
-}
-");
-    }
-
-    [Fact]
-    public async Task ExplicitInterfaceImplementationOptionalParameters_1062_Async()
-    {
-        await TestConversionVisualBasicToCSharpAsync(
-            @"
-Public Interface InterfaceWithOptionalParameters
-    Sub S(Optional i As Integer = 0)
-End Interface
-
-Public Class ImplInterfaceWithOptionalParameters : Implements InterfaceWithOptionalParameters
-    Public Sub InterfaceWithOptionalParameters_S(Optional i As Integer = 0) Implements InterfaceWithOptionalParameters.S
-    End Sub
-End Class", @"
-public partial interface InterfaceWithOptionalParameters
-{
-    void S(int i = 0);
-}
-
-public partial class ImplInterfaceWithOptionalParameters : InterfaceWithOptionalParameters
-{
-    public void InterfaceWithOptionalParameters_S(int i = 0)
-    {
-    }
-
-    void InterfaceWithOptionalParameters.S(int i = 0) => InterfaceWithOptionalParameters_S(i);
-}
-");
     }
 
     [Fact]
@@ -2427,1448 +901,6 @@ public partial class WithOptionalParameters
     {
         S1(@default: ""a"");
     }
-}
-");
-    }
-
-
-    [Fact]
-    public async Task ExplicitInterfaceImplementationOptionalParametersAsync()
-    {
-        await TestConversionVisualBasicToCSharpAsync(
-            @"Public Interface IFoo
-  Property ExplicitProp(Optional str As String = """") As Integer
-  Function ExplicitFunc(Optional str2 As String = """", Optional i2 As Integer = 1) As Integer
-End Interface
-
-Public Class Foo
-  Implements IFoo
-
-  Private Function ExplicitFunc(Optional str As String = """", Optional i2 As Integer = 1) As Integer Implements IFoo.ExplicitFunc
-    Return 5
-  End Function
-    
-  Private Property ExplicitProp(Optional str As String = """") As Integer Implements IFoo.ExplicitProp
-    Get
-      Return 5
-    End Get
-    Set(value As Integer)
-    End Set
-  End Property
-End Class", @"
-public partial interface IFoo
-{
-    int get_ExplicitProp(string str = """");
-    void set_ExplicitProp(string str = """", int value = default);
-    int ExplicitFunc(string str2 = """", int i2 = 1);
-}
-
-public partial class Foo : IFoo
-{
-
-    private int ExplicitFunc(string str = """", int i2 = 1)
-    {
-        return 5;
-    }
-
-    int IFoo.ExplicitFunc(string str = """", int i2 = 1) => ExplicitFunc(str, i2);
-
-    private int get_ExplicitProp(string str = """")
-    {
-        return 5;
-    }
-    private void set_ExplicitProp(string str = """", int value = default)
-    {
-    }
-
-    int IFoo.get_ExplicitProp(string str = """") => get_ExplicitProp(str);
-    void IFoo.set_ExplicitProp(string str = """", int value = default) => set_ExplicitProp(str, value);
-}
-");
-    }
-
-
-    [Fact]
-    public async Task ExplicitInterfaceImplementationOptionalMethodParameters_749_Async()
-    {
-        await TestConversionVisualBasicToCSharpAsync(
-            @"
-Public Interface IFoo
-  Function DoFooBar(ByRef str As String, Optional i As Integer = 4) As Integer
-End Interface
-
-Public Interface IBar
-  Function DoFooBar(ByRef str As String, Optional i As Integer = 8) As Integer
-End Interface
-
-Public Class FooBar
-  Implements IFoo, IBar
-
-  Function Foo(ByRef str As String, Optional i As Integer = 4) As Integer Implements IFoo.DoFooBar
-    Return 4
-  End Function
-
-  Function Bar(ByRef str As String, Optional i As Integer = 8) As Integer Implements IBar.DoFooBar
-    Return 2
-  End Function
-
-End Class", @"
-public partial interface IFoo
-{
-    int DoFooBar(ref string str, int i = 4);
-}
-
-public partial interface IBar
-{
-    int DoFooBar(ref string str, int i = 8);
-}
-
-public partial class FooBar : IFoo, IBar
-{
-
-    public int Foo(ref string str, int i = 4)
-    {
-        return 4;
-    }
-
-    int IFoo.DoFooBar(ref string str, int i = 4) => Foo(ref str, i);
-
-    public int Bar(ref string str, int i = 8)
-    {
-        return 2;
-    }
-
-    int IBar.DoFooBar(ref string str, int i = 8) => Bar(ref str, i);
-
-}
-");
-    }
-
-    [Fact]
-    public async Task RenamedInterfaceMethodFullyQualifiedAsync()
-    {
-        await TestConversionVisualBasicToCSharpAsync(@"Namespace TestNamespace
-    Public Interface IFoo
-        Function DoFoo(ByRef str As String, i As Integer) As Integer
-    End Interface
-End Namespace
-
-Public Class Foo
-    Implements TestNamespace.IFoo
-
-    Function DoFooRenamed(ByRef str As String, i As Integer) As Integer Implements TestNamespace.IFoo.DoFoo
-        Return 4
-    End Function
-End Class", @"
-namespace TestNamespace
-{
-    public partial interface IFoo
-    {
-        int DoFoo(ref string str, int i);
-    }
-}
-
-public partial class Foo : TestNamespace.IFoo
-{
-
-    public int DoFooRenamed(ref string str, int i)
-    {
-        return 4;
-    }
-
-    int TestNamespace.IFoo.DoFoo(ref string str, int i) => DoFooRenamed(ref str, i);
-}");
-    }
-
-    [Fact]
-    public async Task RenamedInterfacePropertyFullyQualifiedAsync()
-    {
-        await TestConversionVisualBasicToCSharpAsync(@"Namespace TestNamespace
-    Public Interface IFoo
-        Property FooProp As Integer
-    End Interface
-End Namespace
-
-Public Class Foo
-    Implements TestNamespace.IFoo
-
-    Property FooPropRenamed As Integer Implements TestNamespace.IFoo.FooProp
-    
-End Class", @"
-namespace TestNamespace
-{
-    public partial interface IFoo
-    {
-        int FooProp { get; set; }
-    }
-}
-
-public partial class Foo : TestNamespace.IFoo
-{
-
-    public int FooPropRenamed { get; set; }
-    int TestNamespace.IFoo.FooProp { get => FooPropRenamed; set => FooPropRenamed = value; }
-
-}");
-    }
-
-    [Fact]
-    public async Task RenamedInterfaceMethodConsumerCasingRenamedAsync()
-    {
-        await TestConversionVisualBasicToCSharpAsync(@"Public Interface IFoo
-        Function DoFoo(ByRef str As String, i As Integer) As Integer
-    End Interface
-
-Public Class Foo
-    Implements IFoo
-
-    Function DoFooRenamed(ByRef str As String, i As Integer) As Integer Implements IFoo.DoFoo
-        Return 4
-    End Function
-End Class
-
-Public Class FooConsumer
-    Function DoFooRenamedConsumer(ByRef str As String, i As Integer) As Integer
-        Dim foo As New Foo
-        Dim bar As IFoo = foo
-        Return foo.DOFOORENAMED(str, i) + bar.DoFoo(str, i)
-    End Function
-End Class", @"
-public partial interface IFoo
-{
-    int DoFoo(ref string str, int i);
-}
-
-public partial class Foo : IFoo
-{
-
-    public int DoFooRenamed(ref string str, int i)
-    {
-        return 4;
-    }
-
-    int IFoo.DoFoo(ref string str, int i) => DoFooRenamed(ref str, i);
-}
-
-public partial class FooConsumer
-{
-    public int DoFooRenamedConsumer(ref string str, int i)
-    {
-        var foo = new Foo();
-        IFoo bar = foo;
-        return foo.DoFooRenamed(ref str, i) + bar.DoFoo(ref str, i);
-    }
-}");
-    }
-
-    [Fact]
-    public async Task RenamedInterfacePropertyConsumerCasingRenamedAsync()
-    {
-        await TestConversionVisualBasicToCSharpAsync(@"Public Interface IFoo
-        Property FooProp As Integer
-    End Interface
-
-Public Class Foo
-    Implements IFoo
-
-    Property FooPropRenamed As Integer Implements IFoo.FooProp
-    
-End Class
-
-Public Class FooConsumer
-    Function GetFooRenamed() As Integer
-        Dim foo As New Foo
-        Dim bar As IFoo = foo
-        Return foo.FOOPROPRENAMED + bar.FooProp
-    End Function
-End Class", @"
-public partial interface IFoo
-{
-    int FooProp { get; set; }
-}
-
-public partial class Foo : IFoo
-{
-
-    public int FooPropRenamed { get; set; }
-    int IFoo.FooProp { get => FooPropRenamed; set => FooPropRenamed = value; }
-
-}
-
-public partial class FooConsumer
-{
-    public int GetFooRenamed()
-    {
-        var foo = new Foo();
-        IFoo bar = foo;
-        return foo.FooPropRenamed + bar.FooProp;
-    }
-}");
-    }
-
-    [Fact]
-    public async Task InterfaceMethodCasingRenamedConsumerAsync()
-    {
-        await TestConversionVisualBasicToCSharpAsync(@"Public Interface IFoo
-        Function DoFoo(str As String, i As Integer) As Integer
-    End Interface
-
-Public Class Foo
-    Implements IFoo
-
-    Function dofoo(str As String, i As Integer) As Integer Implements IFoo.DoFoo
-        Return 4
-    End Function
-End Class
-
-Public Class FooConsumer
-    Function DoFooRenamedConsumer(str As String, i As Integer) As Integer
-        Dim foo As New Foo
-        Dim bar As IFoo = foo
-        Return foo.dofoo(str, i) + bar.DoFoo(str, i)
-    End Function
-End Class", @"
-public partial interface IFoo
-{
-    int DoFoo(string str, int i);
-}
-
-public partial class Foo : IFoo
-{
-
-    public int DoFoo(string str, int i)
-    {
-        return 4;
-    }
-}
-
-public partial class FooConsumer
-{
-    public int DoFooRenamedConsumer(string str, int i)
-    {
-        var foo = new Foo();
-        IFoo bar = foo;
-        return foo.DoFoo(str, i) + bar.DoFoo(str, i);
-    }
-}");
-    }
-
-    [Fact]
-    public async Task InterfacePropertyCasingRenamedConsumerAsync()
-    {
-        await TestConversionVisualBasicToCSharpAsync(@"Public Interface IFoo
-        Property FooProp As Integer
-    End Interface
-
-Public Class Foo
-    Implements IFoo
-
-    Property fooprop As Integer Implements IFoo.FooProp
-    
-End Class
-
-Public Class FooConsumer
-    Function GetFooRenamed() As Integer
-        Dim foo As New Foo
-        Dim bar As IFoo = foo
-        Return foo.fooprop + bar.FooProp
-    End Function
-End Class", @"
-public partial interface IFoo
-{
-    int FooProp { get; set; }
-}
-
-public partial class Foo : IFoo
-{
-
-    public int FooProp { get; set; }
-
-}
-
-public partial class FooConsumer
-{
-    public int GetFooRenamed()
-    {
-        var foo = new Foo();
-        IFoo bar = foo;
-        return foo.FooProp + bar.FooProp;
-    }
-}");
-    }
-
-    [Fact]
-    public async Task InterfaceRenamedMethodConsumerAsync()
-    {
-        await TestConversionVisualBasicToCSharpAsync(@"Public Interface IFoo
-        Function DoFoo(ByRef str As String, i As Integer) As Integer
-    End Interface
-
-Public Class Foo
-    Implements IFoo
-
-    Function DoFooRenamed(ByRef str As String, i As Integer) As Integer Implements IFoo.DoFoo
-        Return 4
-    End Function
-End Class
-
-Public Class FooConsumer
-    Function DoFooRenamedConsumer(ByRef str As String, i As Integer) As Integer
-        Dim foo As New Foo
-        Dim bar As IFoo = foo
-        Return foo.DoFooRenamed(str, i) + bar.DoFoo(str, i)
-    End Function
-End Class", @"
-public partial interface IFoo
-{
-    int DoFoo(ref string str, int i);
-}
-
-public partial class Foo : IFoo
-{
-
-    public int DoFooRenamed(ref string str, int i)
-    {
-        return 4;
-    }
-
-    int IFoo.DoFoo(ref string str, int i) => DoFooRenamed(ref str, i);
-}
-
-public partial class FooConsumer
-{
-    public int DoFooRenamedConsumer(ref string str, int i)
-    {
-        var foo = new Foo();
-        IFoo bar = foo;
-        return foo.DoFooRenamed(ref str, i) + bar.DoFoo(ref str, i);
-    }
-}");
-    }
-
-    [Fact]
-    public async Task InterfaceRenamedPropertyConsumerAsync()
-    {
-        await TestConversionVisualBasicToCSharpAsync(@"Public Interface IFoo
-        Property FooProp As Integer
-    End Interface
-
-Public Class Foo
-    Implements IFoo
-
-    Property FooPropRenamed As Integer Implements IFoo.FooProp
-    
-End Class
-
-Public Class FooConsumer
-    Function GetFooRenamed() As Integer
-        Dim foo As New Foo
-        Dim bar As IFoo = foo
-        Return foo.FooPropRenamed + bar.FooProp
-    End Function
-End Class", @"
-public partial interface IFoo
-{
-    int FooProp { get; set; }
-}
-
-public partial class Foo : IFoo
-{
-
-    public int FooPropRenamed { get; set; }
-    int IFoo.FooProp { get => FooPropRenamed; set => FooPropRenamed = value; }
-
-}
-
-public partial class FooConsumer
-{
-    public int GetFooRenamed()
-    {
-        var foo = new Foo();
-        IFoo bar = foo;
-        return foo.FooPropRenamed + bar.FooProp;
-    }
-}");
-    }
-
-    [Fact]
-    public async Task PartialInterfaceRenamedMethodConsumerAsync()
-    {
-        await TestConversionVisualBasicToCSharpAsync(@"Public Partial Interface IFoo
-        Function DoFoo(ByRef str As String, i As Integer) As Integer
-    End Interface
-
-Public Class Foo
-    Implements IFoo
-
-    Function DoFooRenamed(ByRef str As String, i As Integer) As Integer Implements IFoo.DoFoo
-        Return 4
-    End Function
-End Class
-
-Public Class FooConsumer
-    Function DoFooRenamedConsumer(ByRef str As String, i As Integer) As Integer
-        Dim foo As New Foo
-        Dim bar As IFoo = foo
-        Return foo.DoFooRenamed(str, i) + bar.DoFoo(str, i)
-    End Function
-End Class", @"
-public partial interface IFoo
-{
-    int DoFoo(ref string str, int i);
-}
-
-public partial class Foo : IFoo
-{
-
-    public int DoFooRenamed(ref string str, int i)
-    {
-        return 4;
-    }
-
-    int IFoo.DoFoo(ref string str, int i) => DoFooRenamed(ref str, i);
-}
-
-public partial class FooConsumer
-{
-    public int DoFooRenamedConsumer(ref string str, int i)
-    {
-        var foo = new Foo();
-        IFoo bar = foo;
-        return foo.DoFooRenamed(ref str, i) + bar.DoFoo(ref str, i);
-    }
-}");
-    }
-
-    [Fact]
-    public async Task PartialInterfaceRenamedPropertyConsumerAsync()
-    {
-        await TestConversionVisualBasicToCSharpAsync(@"Public Partial Interface IFoo
-        Property FooProp As Integer
-    End Interface
-
-Public Class Foo
-    Implements IFoo
-
-    Property FooPropRenamed As Integer Implements IFoo.FooProp
-    
-End Class
-
-Public Class FooConsumer
-    Function GetFooRenamed() As Integer
-        Dim foo As New Foo
-        Dim bar As IFoo = foo
-        Return foo.FooPropRenamed + bar.FooProp
-    End Function
-End Class", @"
-public partial interface IFoo
-{
-    int FooProp { get; set; }
-}
-
-public partial class Foo : IFoo
-{
-
-    public int FooPropRenamed { get; set; }
-    int IFoo.FooProp { get => FooPropRenamed; set => FooPropRenamed = value; }
-
-}
-
-public partial class FooConsumer
-{
-    public int GetFooRenamed()
-    {
-        var foo = new Foo();
-        IFoo bar = foo;
-        return foo.FooPropRenamed + bar.FooProp;
-    }
-}");
-    }
-
-    [Fact]
-    public async Task RenamedInterfaceMethodMyClassConsumerAsync()
-    {
-        await TestConversionVisualBasicToCSharpAsync(@"Public Interface IFoo
-        Function DoFoo(ByRef str As String, i As Integer) As Integer
-    End Interface
-
-Public Class Foo
-    Implements IFoo
-
-    Overridable Function DoFooRenamed(ByRef str As String, i As Integer) As Integer Implements IFoo.DoFoo ' Comment ends up out of order, but attached to correct method
-        Return 4
-    End Function
-
-    Function DoFooRenamedConsumer(ByRef str As String, i As Integer) As Integer
-        Return MyClass.DoFooRenamed(str, i)
-    End Function
-End Class", @"
-public partial interface IFoo
-{
-    int DoFoo(ref string str, int i);
-}
-
-public partial class Foo : IFoo
-{
-
-    public int MyClassDoFooRenamed(ref string str, int i)
-    {
-        return 4;
-    }
-
-    int IFoo.DoFoo(ref string str, int i) => DoFooRenamed(ref str, i);
-    public virtual int DoFooRenamed(ref string str, int i) => MyClassDoFooRenamed(ref str, i); // Comment ends up out of order, but attached to correct method
-
-    public int DoFooRenamedConsumer(ref string str, int i)
-    {
-        return MyClassDoFooRenamed(ref str, i);
-    }
-}");
-    }
-
-    [Fact]
-    public async Task RenamedInterfacePropertyMyClassConsumerAsync()
-    {
-        await TestConversionVisualBasicToCSharpAsync(@"Public Interface IFoo
-        ReadOnly Property DoFoo As Integer
-        WriteOnly Property DoBar As Integer
-    End Interface
-
-Public Class Foo
-    Implements IFoo
-
-    Overridable ReadOnly Property DoFooRenamed As Integer Implements IFoo.DoFoo  ' Comment ends up out of order, but attached to correct method
-        Get
-            Return 4
-        End Get
-    End Property
-
-    Overridable WriteOnly Property DoBarRenamed As Integer Implements IFoo.DoBar  ' Comment ends up out of order, but attached to correct method
-        Set
-            Throw New Exception()
-        End Set
-    End Property
-
-    Sub DoFooRenamedConsumer()
-        MyClass.DoBarRenamed = MyClass.DoFooRenamed
-    End Sub
-End Class", @"using System;
-
-public partial interface IFoo
-{
-    int DoFoo { get; }
-    int DoBar { set; }
-}
-
-public partial class Foo : IFoo
-{
-
-    public int MyClassDoFooRenamed
-    {
-        get
-        {
-            return 4;
-        }
-    }
-
-    int IFoo.DoFoo { get => DoFooRenamed; }
-
-    public virtual int DoFooRenamed  // Comment ends up out of order, but attached to correct method
-    {
-        get
-        {
-            return MyClassDoFooRenamed;
-        }
-    }
-
-    public int MyClassDoBarRenamed
-    {
-        set
-        {
-            throw new Exception();
-        }
-    }
-
-    int IFoo.DoBar { set => DoBarRenamed = value; }
-
-    public virtual int DoBarRenamed  // Comment ends up out of order, but attached to correct method
-    {
-        set
-        {
-            MyClassDoBarRenamed = value;
-        }
-    }
-
-    public void DoFooRenamedConsumer()
-    {
-        MyClassDoBarRenamed = MyClassDoFooRenamed;
-    }
-}");
-    }
-
-    [Fact]
-    public async Task ExplicitInterfaceImplementationAsync()
-    {
-        await TestConversionVisualBasicToCSharpAsync(
-            @"Public Interface IFoo
-    Property ExplicitProp(str As String) As Integer
-    Function ExplicitFunc(ByRef str2 As String, i2 As Integer) As Integer
-End Interface
-
-Public Class Foo
-    Implements IFoo
-
-    Private Function ExplicitFunc(ByRef str As String, i As Integer) As Integer Implements IFoo.ExplicitFunc
-        Return 5
-    End Function
-    
-    Private Property ExplicitProp(str As String) As Integer Implements IFoo.ExplicitProp
-        Get
-            Return 5
-        End Get
-        Set(value As Integer)
-        End Set
-    End Property
-End Class", @"
-public partial interface IFoo
-{
-    int get_ExplicitProp(string str);
-    void set_ExplicitProp(string str, int value);
-    int ExplicitFunc(ref string str2, int i2);
-}
-
-public partial class Foo : IFoo
-{
-
-    private int ExplicitFunc(ref string str, int i)
-    {
-        return 5;
-    }
-
-    int IFoo.ExplicitFunc(ref string str, int i) => ExplicitFunc(ref str, i);
-
-    private int get_ExplicitProp(string str)
-    {
-        return 5;
-    }
-    private void set_ExplicitProp(string str, int value)
-    {
-    }
-
-    int IFoo.get_ExplicitProp(string str) => get_ExplicitProp(str);
-    void IFoo.set_ExplicitProp(string str, int value) => set_ExplicitProp(str, value);
-}
-");
-    }
-
-    [Fact]
-    public async Task PropertyInterfaceImplementationKeepsVirtualModifierAsync()
-    {
-        await TestConversionVisualBasicToCSharpAsync(
-            @"Public Interface IFoo
-    Property PropParams(str As String) As Integer
-    Property Prop() As Integer
-End Interface
-
-Public Class Foo
-    Implements IFoo
-
-    Public Overridable Property PropParams(str As String) As Integer Implements IFoo.PropParams
-        Get
-            Return 5
-        End Get
-        Set(value As Integer)
-        End Set
-    End Property
-
-    Public Overridable Property Prop As Integer Implements IFoo.Prop
-        Get
-            Return 5
-        End Get
-        Set(value As Integer)
-        End Set
-    End Property
-End Class", @"
-public partial interface IFoo
-{
-    int get_PropParams(string str);
-    void set_PropParams(string str, int value);
-    int Prop { get; set; }
-}
-
-public partial class Foo : IFoo
-{
-
-    public virtual int get_PropParams(string str)
-    {
-        return 5;
-    }
-    public virtual void set_PropParams(string str, int value)
-    {
-    }
-
-    public virtual int Prop
-    {
-        get
-        {
-            return 5;
-        }
-        set
-        {
-        }
-    }
-}
-");
-    }
-
-    [Fact]
-    public async Task PrivateAutoPropertyImplementsMultipleInterfacesAsync()
-    {
-        await TestConversionVisualBasicToCSharpAsync(
-            @"Public Interface IFoo
-    Property ExplicitProp As Integer
-End Interface
-
-Public Interface IBar
-    Property ExplicitProp As Integer
-End Interface
-
-Public Class Foo
-    Implements IFoo, IBar
-    
-    Private Property ExplicitProp As Integer Implements IFoo.ExplicitProp, IBar.ExplicitProp
-End Class", @"
-public partial interface IFoo
-{
-    int ExplicitProp { get; set; }
-}
-
-public partial interface IBar
-{
-    int ExplicitProp { get; set; }
-}
-
-public partial class Foo : IFoo, IBar
-{
-
-    private int ExplicitProp { get; set; }
-    int IFoo.ExplicitProp { get => ExplicitProp; set => ExplicitProp = value; }
-    int IBar.ExplicitProp { get => ExplicitProp; set => ExplicitProp = value; }
-}");
-    }
-
-        
-    [Fact]
-    public async Task ImplementMultipleRenamedPropertiesFromInterfaceAsAbstractAsync()
-    {
-        await TestConversionVisualBasicToCSharpAsync(
-            @"
-Public Interface IFoo
-    Property ExplicitProp As Integer
-End Interface
-Public Interface IBar
-    Property ExplicitProp As Integer
-End Interface
-Public MustInherit Class Foo
-    Implements IFoo, IBar
-
-    Protected MustOverride Property ExplicitPropRenamed1 As Integer Implements IFoo.ExplicitProp
-    Protected MustOverride Property ExplicitPropRenamed2 As Integer Implements IBar.ExplicitProp
-End Class", @"
-public partial interface IFoo
-{
-    int ExplicitProp { get; set; }
-}
-
-public partial interface IBar
-{
-    int ExplicitProp { get; set; }
-}
-public abstract partial class Foo : IFoo, IBar
-{
-
-    protected abstract int ExplicitPropRenamed1 { get; set; }
-    int IFoo.ExplicitProp { get => ExplicitPropRenamed1; set => ExplicitPropRenamed1 = value; }
-    protected abstract int ExplicitPropRenamed2 { get; set; }
-    int IBar.ExplicitProp { get => ExplicitPropRenamed2; set => ExplicitPropRenamed2 = value; }
-}");
-    }
-
-    [Fact]
-    public async Task ExplicitInterfaceImplementationForVirtualMemberFromAnotherClassAsync()
-    {
-        await TestConversionVisualBasicToCSharpAsync(
-            @"
-Public Interface IFoo
-    Sub Save()
-    Property Prop As Integer
-End Interface
-
-Public MustInherit Class BaseFoo
-    Protected Overridable Sub OnSave()
-    End Sub
-
-    Protected Overridable Property MyProp As Integer = 5
-End Class
-
-Public Class Foo
-    Inherits BaseFoo
-    Implements IFoo
-
-    Protected Overrides Sub OnSave() Implements IFoo.Save
-    End Sub
-
-    Protected Overrides Property MyProp As Integer = 6 Implements IFoo.Prop
-
-End Class", @"
-public partial interface IFoo
-{
-    void Save();
-    int Prop { get; set; }
-}
-
-public abstract partial class BaseFoo
-{
-    protected virtual void OnSave()
-    {
-    }
-
-    protected virtual int MyProp { get; set; } = 5;
-}
-
-public partial class Foo : BaseFoo, IFoo
-{
-
-    protected override void OnSave()
-    {
-    }
-
-    void IFoo.Save() => OnSave();
-
-    protected override int MyProp { get; set; } = 6;
-    int IFoo.Prop { get => MyProp; set => MyProp = value; }
-
-}");
-    }
-
-    [Fact]
-    public async Task ExplicitInterfaceImplementationWhereOnlyOneInterfaceMemberIsRenamedAsync()
-    {
-        await TestConversionVisualBasicToCSharpAsync(
-            @"
-Public Interface IFoo
-    Sub Save()
-    Property A As Integer
-End Interface
-
-Public Interface IBar
-    Sub OnSave()
-    Property B As Integer
-End Interface
-
-Public Class Foo
-    Implements IFoo, IBar
-
-    Public Overridable Sub Save() Implements IFoo.Save, IBar.OnSave
-    End Sub
-
-    Public Overridable Property A As Integer Implements IFoo.A, IBar.B
-
-End Class", @"
-public partial interface IFoo
-{
-    void Save();
-    int A { get; set; }
-}
-
-public partial interface IBar
-{
-    void OnSave();
-    int B { get; set; }
-}
-
-public partial class Foo : IFoo, IBar
-{
-
-    public virtual void Save()
-    {
-    }
-
-    void IFoo.Save() => Save();
-    void IBar.OnSave() => Save();
-
-    public virtual int A { get; set; }
-    int IFoo.A { get => A; set => A = value; }
-    int IBar.B { get => A; set => A = value; }
-
-}");
-    }
-
-    [Fact]
-    public async Task ExplicitInterfaceImplementationWhereMemberShadowsBaseAsync()
-    {
-        await TestConversionVisualBasicToCSharpAsync(
-            @"
-Public Interface IFoo
-    Sub Save()
-    Property Prop As Integer
-End Interface
-
-Public MustInherit Class BaseFoo
-    Public Overridable Sub OnSave()
-    End Sub
-
-    Public Overridable Property MyProp As Integer = 5
-End Class
-
-Public Class Foo
-    Inherits BaseFoo
-    Implements IFoo
-
-    Public Shadows Sub OnSave() Implements IFoo.Save
-    End Sub
-
-    Public Shadows Property MyProp As Integer = 6 Implements IFoo.Prop
-
-End Class", @"
-public partial interface IFoo
-{
-    void Save();
-    int Prop { get; set; }
-}
-
-public abstract partial class BaseFoo
-{
-    public virtual void OnSave()
-    {
-    }
-
-    public virtual int MyProp { get; set; } = 5;
-}
-
-public partial class Foo : BaseFoo, IFoo
-{
-
-    public new void OnSave()
-    {
-    }
-
-    void IFoo.Save() => OnSave();
-
-    public new int MyProp { get; set; } = 6;
-    int IFoo.Prop { get => MyProp; set => MyProp = value; }
-
-}");
-    }
-
-    [Fact]
-    public async Task PrivatePropertyAccessorBlocksImplementsMultipleInterfacesAsync()
-    {
-        await TestConversionVisualBasicToCSharpAsync(
-            @"Public Interface IFoo
-    Property ExplicitProp As Integer
-End Interface
-
-Public Interface IBar
-    Property ExplicitProp As Integer
-End Interface
-
-Public Class Foo
-    Implements IFoo, IBar
-    
-    Private Property ExplicitProp As Integer Implements IFoo.ExplicitProp, IBar.ExplicitProp ' Comment moves because this line gets split
-        Get
-          Return 5
-        End Get
-        Set
-        End Set
-    End Property
-End Class", @"
-public partial interface IFoo
-{
-    int ExplicitProp { get; set; }
-}
-
-public partial interface IBar
-{
-    int ExplicitProp { get; set; }
-}
-
-public partial class Foo : IFoo, IBar
-{
-
-    private int ExplicitProp
-    {
-        get
-        {
-            return 5;
-        }
-        set
-        {
-        }
-    }
-
-    int IFoo.ExplicitProp { get => ExplicitProp; set => ExplicitProp = value; }
-    int IBar.ExplicitProp { get => ExplicitProp; set => ExplicitProp = value; } // Comment moves because this line gets split
-}");
-    }
-
-    [Fact]
-    public async Task NonPublicImplementsInterfacesAsync()
-    {
-        await TestConversionVisualBasicToCSharpAsync(
-            @"Public Interface IFoo
-    Property FriendProp As Integer
-    Sub ProtectedSub()
-    Function PrivateFunc() As Integer
-    Sub ProtectedInternalSub()
-    Sub AbstractSub()
-End Interface
-
-Public Interface IBar
-    Property FriendProp As Integer
-    Sub ProtectedSub()
-    Function PrivateFunc() As Integer
-    Sub ProtectedInternalSub()
-    Sub AbstractSub()
-End Interface
-
-Public MustInherit Class BaseFoo
-    Implements IFoo, IBar
-    
-    Friend Overridable Property FriendProp As Integer Implements IFoo.FriendProp, IBar.FriendProp ' Comment moves because this line gets split
-        Get
-          Return 5
-        End Get
-        Set
-        End Set
-    End Property
-
-    Protected Sub ProtectedSub() Implements IFoo.ProtectedSub, IBar.ProtectedSub
-    End Sub
-
-    Private Function PrivateFunc() As Integer Implements IFoo.PrivateFunc, IBar.PrivateFunc
-    End Function
-
-    Protected Friend Overridable Sub ProtectedInternalSub() Implements IFoo.ProtectedInternalSub, IBar.ProtectedInternalSub
-    End Sub
-
-    Protected MustOverride Sub AbstractSubRenamed() Implements IFoo.AbstractSub, IBar.AbstractSub
-End Class
-
-Public Class Foo
-    Inherits BaseFoo
-
-    Protected Friend Overrides Sub ProtectedInternalSub()
-    End Sub
-
-    Protected Overrides Sub AbstractSubRenamed()
-    End Sub
-End Class
-", @"
-public partial interface IFoo
-{
-    int FriendProp { get; set; }
-    void ProtectedSub();
-    int PrivateFunc();
-    void ProtectedInternalSub();
-    void AbstractSub();
-}
-
-public partial interface IBar
-{
-    int FriendProp { get; set; }
-    void ProtectedSub();
-    int PrivateFunc();
-    void ProtectedInternalSub();
-    void AbstractSub();
-}
-
-public abstract partial class BaseFoo : IFoo, IBar
-{
-
-    internal virtual int FriendProp
-    {
-        get
-        {
-            return 5;
-        }
-        set
-        {
-        }
-    }
-
-    int IFoo.FriendProp { get => FriendProp; set => FriendProp = value; }
-    int IBar.FriendProp { get => FriendProp; set => FriendProp = value; } // Comment moves because this line gets split
-
-    protected void ProtectedSub()
-    {
-    }
-
-    void IFoo.ProtectedSub() => ProtectedSub();
-    void IBar.ProtectedSub() => ProtectedSub();
-
-    private int PrivateFunc()
-    {
-        return default;
-    }
-
-    int IFoo.PrivateFunc() => PrivateFunc();
-    int IBar.PrivateFunc() => PrivateFunc();
-
-    protected internal virtual void ProtectedInternalSub()
-    {
-    }
-
-    void IFoo.ProtectedInternalSub() => ProtectedInternalSub();
-    void IBar.ProtectedInternalSub() => ProtectedInternalSub();
-
-    protected abstract void AbstractSubRenamed();
-    void IFoo.AbstractSub() => AbstractSubRenamed();
-    void IBar.AbstractSub() => AbstractSubRenamed();
-}
-
-public partial class Foo : BaseFoo
-{
-
-    protected internal override void ProtectedInternalSub()
-    {
-    }
-
-    protected override void AbstractSubRenamed()
-    {
-    }
-}");
-    }
-
-    [Fact]
-    public async Task ExplicitPropertyImplementationWithDirectAccessAsync()
-    {
-        await TestConversionVisualBasicToCSharpAsync(
-            @"
-Public Interface IFoo
-    Property ExplicitProp As Integer
-    ReadOnly Property ExplicitReadOnlyProp As Integer
-End Interface
-
-Public Class Foo
-    Implements IFoo
-    
-    Property ExplicitPropRenamed As Integer Implements IFoo.ExplicitProp
-    ReadOnly Property ExplicitRenamedReadOnlyProp As Integer Implements IFoo.ExplicitReadOnlyProp
-
-    Private Sub Consumer()
-        _ExplicitPropRenamed = 5
-        _ExplicitRenamedReadOnlyProp = 10
-    End Sub
-
-End Class", @"
-public partial interface IFoo
-{
-    int ExplicitProp { get; set; }
-    int ExplicitReadOnlyProp { get; }
-}
-
-public partial class Foo : IFoo
-{
-
-    public int ExplicitPropRenamed { get; set; }
-    int IFoo.ExplicitProp { get => ExplicitPropRenamed; set => ExplicitPropRenamed = value; }
-    public int ExplicitRenamedReadOnlyProp { get; private set; }
-    int IFoo.ExplicitReadOnlyProp { get => ExplicitRenamedReadOnlyProp; }
-
-    private void Consumer()
-    {
-        ExplicitPropRenamed = 5;
-        ExplicitRenamedReadOnlyProp = 10;
-    }
-
-}");
-    }
-        
-    [Fact]
-    public async Task ReadonlyRenamedPropertyImplementsMultipleInterfacesAsync()
-    {
-        await TestConversionVisualBasicToCSharpAsync(
-            @"Public Interface IFoo
-    ReadOnly Property ExplicitProp As Integer
-End Interface
-
-Public Interface IBar
-    ReadOnly Property ExplicitProp As Integer
-End Interface
-
-Public Class Foo
-    Implements IFoo, IBar
-    
-    ReadOnly Property ExplicitPropRenamed As Integer Implements IFoo.ExplicitProp, IBar.ExplicitProp
-End Class", @"
-public partial interface IFoo
-{
-    int ExplicitProp { get; }
-}
-
-public partial interface IBar
-{
-    int ExplicitProp { get; }
-}
-
-public partial class Foo : IFoo, IBar
-{
-
-    public int ExplicitPropRenamed { get; private set; }
-    int IFoo.ExplicitProp { get => ExplicitPropRenamed; }
-    int IBar.ExplicitProp { get => ExplicitPropRenamed; }
-}");
-    }
-
-    [Fact]
-    public async Task WriteonlyPropertyImplementsMultipleInterfacesAsync()
-    {
-        await TestConversionVisualBasicToCSharpAsync(
-            @"Public Interface IFoo
-    WriteOnly Property ExplicitProp As Integer
-End Interface
-
-Public Interface IBar
-    WriteOnly Property ExplicitProp As Integer
-End Interface
-
-Public Class Foo
-    Implements IFoo, IBar
-    
-    WriteOnly Property ExplicitPropRenamed As Integer Implements IFoo.ExplicitProp, IBar.ExplicitProp ' Comment moves because this line gets split
-        Set
-        End Set        
-    End Property
-End Class", @"
-public partial interface IFoo
-{
-    int ExplicitProp { set; }
-}
-
-public partial interface IBar
-{
-    int ExplicitProp { set; }
-}
-
-public partial class Foo : IFoo, IBar
-{
-
-    public int ExplicitPropRenamed
-    {
-        set
-        {
-        }
-    }
-
-    int IFoo.ExplicitProp { set => ExplicitPropRenamed = value; }
-    int IBar.ExplicitProp { set => ExplicitPropRenamed = value; } // Comment moves because this line gets split
-}");
-    }
-
-    [Fact]
-    public async Task PrivateMethodAndParameterizedPropertyImplementsMultipleInterfacesAsync()
-    {
-        await TestConversionVisualBasicToCSharpAsync(
-            @"Public Interface IFoo
-    Property ExplicitProp(str As String) As Integer
-    Function ExplicitFunc(ByRef str2 As String, i2 As Integer) As Integer
-End Interface
-
-Public Interface IBar
-    Property ExplicitProp(str As String) As Integer
-    Function ExplicitFunc(ByRef str2 As String, i2 As Integer) As Integer
-End Interface
-
-Public Class Foo
-    Implements IFoo, IBar
-
-    Private Function ExplicitFunc(ByRef str As String, i As Integer) As Integer Implements IFoo.ExplicitFunc, IBar.ExplicitFunc
-        Return 5
-    End Function
-    
-    Private Property ExplicitProp(str As String) As Integer Implements IFoo.ExplicitProp, IBar.ExplicitProp
-        Get
-            Return 5
-        End Get
-        Set(value As Integer)
-        End Set
-    End Property
-End Class", @"
-public partial interface IFoo
-{
-    int get_ExplicitProp(string str);
-    void set_ExplicitProp(string str, int value);
-    int ExplicitFunc(ref string str2, int i2);
-}
-
-public partial interface IBar
-{
-    int get_ExplicitProp(string str);
-    void set_ExplicitProp(string str, int value);
-    int ExplicitFunc(ref string str2, int i2);
-}
-
-public partial class Foo : IFoo, IBar
-{
-
-    private int ExplicitFunc(ref string str, int i)
-    {
-        return 5;
-    }
-
-    int IFoo.ExplicitFunc(ref string str, int i) => ExplicitFunc(ref str, i);
-    int IBar.ExplicitFunc(ref string str, int i) => ExplicitFunc(ref str, i);
-
-    private int get_ExplicitProp(string str)
-    {
-        return 5;
-    }
-    private void set_ExplicitProp(string str, int value)
-    {
-    }
-
-    int IFoo.get_ExplicitProp(string str) => get_ExplicitProp(str);
-    int IBar.get_ExplicitProp(string str) => get_ExplicitProp(str);
-    void IFoo.set_ExplicitProp(string str, int value) => set_ExplicitProp(str, value);
-    void IBar.set_ExplicitProp(string str, int value) => set_ExplicitProp(str, value);
-}");
-    }
-
-    /// <summary>
-    /// 
-    /// </summary>
-    /// <returns></returns>
-    [Fact]
-    public async Task Issue444_InternalMemberDelegatingMethodAsync()
-    {
-        await TestConversionVisualBasicToCSharpAsync(
-            @"Public Interface IFoo
-    Function FooDifferentName(ByRef str As String, i As Integer) As Integer
-End Interface
-
-Friend Class Foo
-    Implements IFoo
-
-    Function BarDifferentName(ByRef str As String, i As Integer) As Integer Implements IFoo.FooDifferentName
-        Return 4
-    End Function
-End Class", @"
-public partial interface IFoo
-{
-    int FooDifferentName(ref string str, int i);
-}
-
-internal partial class Foo : IFoo
-{
-
-    public int BarDifferentName(ref string str, int i)
-    {
-        return 4;
-    }
-
-    int IFoo.FooDifferentName(ref string str, int i) => BarDifferentName(ref str, i);
 }
 ");
     }
@@ -3920,79 +952,6 @@ public enum MyEnum
     MyEnum
 }
 ");
-    }
-
-    [Fact]
-    public async Task TestConstructorStaticLocalConvertedToFieldAsync()
-    {
-        await TestConversionVisualBasicToCSharpAsync(
-            @"Class StaticLocalConvertedToField
-    Sub New(x As Boolean)
-        Static sPrevPosition As Integer = 7 ' Comment moves with declaration
-        Console.WriteLine(sPrevPosition)
-    End Sub
-    Sub New(x As Integer)
-        Static sPrevPosition As Integer
-        Console.WriteLine(sPrevPosition)
-    End Sub
-End Class", @"using System;
-
-internal partial class StaticLocalConvertedToField
-{
-    private int _sPrevPosition = 7; // Comment moves with declaration
-    public StaticLocalConvertedToField(bool x)
-    {
-        Console.WriteLine(_sPrevPosition);
-    }
-
-    private int _sPrevPosition1 = default;
-    public StaticLocalConvertedToField(int x)
-    {
-        Console.WriteLine(_sPrevPosition1);
-    }
-}");
-    }
-
-    [Fact]
-    public async Task TestPropertyStaticLocalConvertedToFieldAsync()
-    {
-        await TestConversionVisualBasicToCSharpAsync(
-            @"Class StaticLocalConvertedToField
-    Readonly Property OtherName() As Integer
-        Get
-            Static sPrevPosition As Integer = 3 ' Comment moves with declaration
-            Console.WriteLine(sPrevPosition)
-            Return sPrevPosition
-        End Get
-    End Property
-    Readonly Property OtherName(x As Integer) as Integer
-        Get
-            Static sPrevPosition As Integer
-            sPrevPosition += 1
-            Return sPrevPosition
-        End Get
-    End Property
-End Class", @"using System;
-
-internal partial class StaticLocalConvertedToField
-{
-    private int _OtherName_sPrevPosition = 3; // Comment moves with declaration
-    public int OtherName
-    {
-        get
-        {
-            Console.WriteLine(_OtherName_sPrevPosition);
-            return _OtherName_sPrevPosition;
-        }
-    }
-
-    private int _OtherName_sPrevPosition1 = default;
-    public int get_OtherName(int x)
-    {
-        _OtherName_sPrevPosition1 += 1;
-        return _OtherName_sPrevPosition1;
-    }
-}");
     }
 
     [Fact]
